Refactoring clean packages in leshan-client-core
